+++ src/common/enum/MessageBoxResponse.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2018 Damir Porobic <damir.porobic@gmx.com>
+ * Copyright (C) 2019 Damir Porobic <https://github.com/damirporobic>
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as published by
@@ -17,22 +17,14 @@
  * Boston, MA 02110-1301, USA.
  */
 
-#ifndef KSNIP_MESSAGEBOXHELPER_H
-#define KSNIP_MESSAGEBOXHELPER_H
+#ifndef KSNIP_MESSAGEBOXRESPONSE_H
+#define KSNIP_MESSAGEBOXRESPONSE_H
 
-#include <QMessageBox>
-
-#include "src/common/enum/MessageBoxResponse.h"
-
-class MessageBoxHelper
+enum class MessageBoxResponse
 {
-public:
-    static bool yesNo(const QString &title, const QString &question);
-    static MessageBoxResponse yesNoCancel(const QString &title, const QString &question);
-    static void ok(const QString &title, const QString &info);
-
-private:
-	static MessageBoxResponse mapReplyToMessageBoxResponse(int reply);
+	Yes,
+	No,
+	Cancel
 };
 
-#endif //KSNIP_MESSAGEBOXHELPER_H
+#endif //KSNIP_MESSAGEBOXRESPONSE_H